% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absorptionDat}
\alias{absorptionDat}
\title{absorptionDat}
\format{A data frame with 302 obs. of  2 variables:
\describe{
  \item{\code{nm}}{double Wavelength in nm}
  \item{\code{Absorption}}{double Normalised absorption spectra}
}}
\usage{
absorptionDat
}
\description{
Absorption data
}
\keyword{datasets}
