% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actionSpectrumDat}
\alias{actionSpectrumDat}
\title{actionSpectrumDat}
\format{A data frame with 1813 obs. of  3 variables:
\describe{
  \item{\code{study.spectra}}{study.spectra}
  \item{\code{nm}}{nm}
  \item{\code{Absorption}}{Absorption}   
}}
\source{
\url{}
}
\description{
Action spectra data from various sources
}
