% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMO_Light_model.R
\name{calc.PAR}
\alias{calc.PAR}
\title{calc.PAR}
\usage{
calc.PAR(
  SZA = 0,
  TSS = 0,
  Depth = 0,
  solar.zenith.dat,
  model.dat,
  absorption.dat,
  ND.adj = 1,
  nm.range = c(400, 700)
)
}
\arguments{
\item{SZA}{Solar Zenthis Angle}

\item{TSS}{Total suspended solids level in the water column (generally derived as conversion factor from NTU readings)}

\item{Depth}{water depth at which the light reading should be estimated}

\item{solar.zenith.dat}{Fraction of illumination given a solar zenith angle (SZA)}

\item{model.dat}{Model data derived from IMO light measurements}

\item{absorption.dat}{A data.frame containing an absorption spectrum to use for calculating PUR. Must contain columns fo nm and normalized absoption/action (values between zero and one)}

\item{ND.adj}{xxx}

\item{nm.range}{Light eavelength range over which PUR and PAR will be calculated}
}
\value{
Calculated PUR and PAR light data.
}
\description{
Calculates PAR and PUR for a given depth and TSS level, based on solar zenthith angle, observed model data collected in Cleveland Bay and an absportion spectra
}
