% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experimentalSpectrumDat}
\alias{experimentalSpectrumDat}
\title{Experiment spectra data}
\format{A data frame with 1813 obs. of  3 variables:
\describe{
  \item{\code{Light.source}}{Light.source}
  \item{\code{Wavelength}}{Wavelength}
  \item{\code{Relative.Intensity}}{Relative.Intensity}   
}}
\source{
various
}
\usage{
experimentalSpectrumDat
}
\description{
Action spectra data from various sources
}
\examples{
\dontrun{
experimentalSpectrumDat
} 
}
\keyword{datasets}
