% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelDat}
\alias{modelDat}
\title{modelDat}
\format{A data frame with 351 obs. of  7 variables:
\describe{
  \item{\code{nm}}{integer Wavelength in nm}
  \item{\code{Kd_M}}{Kd_M}
  \item{\code{Kd_C}}{Kd_C}
  \item{\code{Es_C0}}{double Es_C0}
  \item{\code{Es_C1}}{double Es_C1}
  \item{\code{Es_C2}}{Es_C2}
  \item{\code{Es_C3}}{Es_C2}
}}
\source{
\url{}
}
\usage{
modelDat
}
\description{
Model data derived from IMO observations in Cleveland Bay
}
\keyword{datasets}
