% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_light_MS8.R
\name{parse.MS8.logger.dat}
\alias{parse.MS8.logger.dat}
\title{parse.MS8.logger.dat}
\usage{
parse.MS8.logger.dat(file.f)
}
\arguments{
\item{file.f}{The MS8 logger file to be processed (including path, if required)}
}
\value{
A list containing the processed MS8 logger data, including a combined dataset as well as the individual channels present.
}
\description{
Reads in MS8 logger data
}
