% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solar.zenith.dat}
\alias{solar.zenith.dat}
\title{solar.zenith.dat}
\format{A data frame with 91 obs. of  2 variables:
\describe{
  \item{\code{sza}}{integer Solar zenith angle}
  \item{\code{frac}}{frac}
}}
\source{
\url{}
}
\usage{
solar.zenith.dat
}
\description{
MSolar zenith data
}
\keyword{datasets}
