% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solarZenithDat}
\alias{solarZenithDat}
\title{solarZenithDat}
\format{A data frame with 91 obs. of  2 variables:
\describe{
  \item{\code{sza}}{integer Solar zenith angle}
  \item{\code{frac}}{frac}
}}
\source{
\url{}
}
\description{
MSolar zenith data
}
